Added more constants, module formulas removed
