+++ tutorlolv2_math/src/__v2/arena/mod.rs
@@ -1,5 +1,4 @@
 pub mod calc;
-pub mod formulas;
-pub mod helpers_arena;
+pub mod helpers;
 pub mod model;
 pub mod rt;
