+++ tutorlolv2_math/src/__v2/arena/model.rs
@@ -1,4 +1,4 @@
-use crate::__v2::{AbilityLevels, L_ITEM, L_RUNE, L_SIML, riot::RiotChampionStats};
+use crate::__v2::{AbilityLevels, L_ITEM, L_RUNE, L_SIML, ResistValue, riot::RiotChampionStats};
 use bincode::{Decode, Encode};
 use smallvec::SmallVec;
 use tinyset::SetU32;
@@ -215,11 +215,6 @@ pub struct EnemyState {
     pub stacks: u32,
     pub champion_id: ChampionId,
     pub level: u8,
-}
-
-pub struct ResistValue {
-    pub real: f32,
-    pub modifier: f32,
 }
 
 pub struct SelfState {
