+++ tutorlolv2_math/src/__v2/arena/rt.rs
@@ -1,5 +1,5 @@
-use super::{formulas::*, helpers_arena::*, model::*};
-use crate::__v2::{L_SIML, riot::*};
+use super::{helpers::*, model::*};
+use crate::__v2::{L_SIML, RiotFormulas, riot::*};
 use std::mem::MaybeUninit;
 use tinyset::SetU32;
 use tutorlolv2_gen::{
