+++ tutorlolv2_math/src/__v2/stdalloc/mod.rs
@@ -1,5 +1,4 @@
 pub mod calc;
-pub mod formulas;
 pub mod helpers;
 pub mod model;
 pub mod rt;
